🚧  AppBar refactor to ensure no mutation of navigation
