+++ sites/jeromefitzgerald.com/src/components/AppBar/MenuKBar.tsx
@@ -1,0 +1,7 @@
+import { KBarPortal } from '~components/KBar'
+
+const MenuKBar = () => {
+  return <KBarPortal />
+}
+
+export { MenuKBar }
